requirements.txt
discord.py>=2.0.0
yfinance>=0.2.0
